Outsource more to utils and add safeguards for filenames with spaces
