+++ clip.js
@@ -1,30 +1,19 @@
-const fs = require('fs');
-const path = require('path');
-const execSync = require('child_process').execSync;
 const parser = require('args-parser');
-
-function extractClip(filePath, start, duration) {
-
-    const fileName = path.basename(filePath);
-    const directory = path.dirname(filePath);
-    const extension = path.extname(filePath);
-
-    const outputFileName = fileName.replace(extension, '') + '_clip' + 
-        `_${start}`.replace('.', '-') + 
-        `_${start + duration}`.replace('.', '-') + 
-        extension;
-
-    const command = `ffmpeg -ss ${start} -i ${fileName} -c copy -t ${duration} ${outputFileName}`;
-    console.log(command);
-    execSync(command, { cwd: directory, stdio: 'inherit' });
-}
+const chalk = require('chalk');
+const { extractClip, fade } = require('./utils');
 
 function run() {
     let args = parser(process.argv);
     let filePath = args.file;
     if (args.start && args.duration) {
-        extractClip(filePath, args.start, args.duration);
+        filePath = extractClip(filePath, args.start, args.duration);
+        console.log(chalk.gray(`Clip extraction result written to ${filePath}`));
+        if (args.fade) {
+            filePath = fade(filePath, args.fade);
+            console.log(chalk.gray(`Fade result written to ${filePath}`));
+        }
     }
+    console.log(chalk.green(`Clip result written to ${filePath}`));
 }
 
 run();
