+++ package.json
@@ -1,5 +1,6 @@
 {
   "dependencies": {
-    "args-parser": "^1.3.0"
+    "args-parser": "^1.3.0",
+    "chalk": "^4.1.2"
   }
 }
