+++ utils.js
@@ -5,9 +5,17 @@ function getDuration(filePath) {
     const directory = path.dirname(filePath);
     const fileName = path.basename(filePath);
     const command = `ffprobe -i ${fileName} -show_entries format=duration -v quiet -of csv="p=0"`;
-    const duration = parseFloat(execSync(command, { cwd: directory }));
-    console.log(`Got duration ${duration} with command : ${command}`);
-    return duration;
+    //const command = `ffprobe`;
+    try {
+        const durationResult = execSync(command, { cwd: directory });
+        const duration = parseFloat(durationResult);
+        console.log(`Got duration ${duration} with command : ${command}`);
+        return duration;
+    } catch(error) {
+        console.error(error);
+        throw error;
+    }
+    
 }
 
 function createGIF(filePath, start, duration, scale) {
@@ -23,7 +31,7 @@ function createGIF(filePath, start, duration, scale) {
 
     const outputFileName = fileName.replace(extension, '') + '_short.gif';
 
-    const command = `ffmpeg -ss ${start} -t ${duration} -i ${fileName} -filter_complex "[0:v] fps=24,scale=${scale}:-1,split [a][b];[a] palettegen [p];[b][p] paletteuse" ${outputFileName}`;
+    const command = `ffmpeg -ss ${start} -t ${duration} -i "${fileName}" -filter_complex "[0:v] fps=24,scale=${scale}:-1,split [a][b];[a] palettegen [p];[b][p] paletteuse" "${outputFileName}"`;
 
     console.log(command);
 
@@ -32,7 +40,87 @@ function createGIF(filePath, start, duration, scale) {
     return path.join(directory, outputFileName);
 }
 
+function extractClip(filePath, start, duration) {
+
+    const fileName = path.basename(filePath);
+    const directory = path.dirname(filePath);
+    const extension = path.extname(filePath);
+
+    const outputFileName = fileName.replace(extension, '') + '_clip' + 
+        `_${start}`.replace('.', '-') + 
+        `_${start + duration}`.replace('.', '-') + 
+        extension;
+
+    const command = `ffmpeg -ss ${start} -i "${fileName}" -c copy -t ${duration} "${outputFileName}"`;
+    console.log(command);
+    execSync(command, { cwd: directory, stdio: 'inherit' });
+
+    return path.join(directory, outputFileName);
+}
+
+function fade(filePath, fadeDuration) {
+
+    // https://dev.to/dak425/add-fade-in-and-fade-out-effects-with-ffmpeg-2bj7
+
+    const fileDuration = getDuration(filePath)
+    const directory = path.dirname(filePath);
+    const fileName = path.basename(filePath);
+    const extension = path.extname(filePath);
+    const outputFileName = fileName.replace(extension, '') + '_faded' + extension;
+
+
+    let command = `ffmpeg -i "${fileName}" `;
+    command += `-vf "fade=t=in:st=0:d=${fadeDuration},fade=t=out:st=${fileDuration-fadeDuration}:d=${fadeDuration}" `;
+    command += `-af "afade=t=in:st=0:d=${fadeDuration},afade=t=out:st=${fileDuration-fadeDuration}:d=${fadeDuration}" `;
+    command += ` "${outputFileName}"`;
+    
+    console.log(command);
+
+    execSync(command, { cwd: directory, stdio: 'inherit' });
+
+    return path.join(directory, outputFileName);
+}
+
+function truncate(filePath, end) {
+
+    const directory = path.dirname(filePath);
+    const fileName = path.basename(filePath);
+    const extension = path.extname(filePath);
+    const outputFileName = fileName.replace(extension, '') + '_truncated' + extension;
+
+    const command = `ffmpeg -i ${fileName} -c copy -t ${end} ${outputFileName}`;
+
+    console.log(command);
+
+    execSync(command, { cwd: directory, stdio: 'inherit' });
+
+    return path.join(directory, outputFileName);
+}
+
+function delayAudio(filePath, seconds) {
+
+    const directory = path.dirname(filePath);
+    const fileName = path.basename(filePath);
+    const extension = path.extname(filePath);
+    const outputFileName = fileName.replace(extension, '') + '_delayed' + extension;
+    
+    const command = `ffmpeg -i "${fileName}" -itsoffset ${seconds} -i "${fileName}" -map 0:v -map 1:a -c copy "${outputFileName}"`;
+
+    // ffmpeg -i "Movie_008.mp4" -itsoffset 0.12 -i "Movie_008.mp4" -map 0:v -map 1:a -c copy "Movie_008_delayed.mp4"
+
+    console.log(command);
+
+    execSync(command, { cwd: directory, stdio: 'inherit' });
+
+    return path.join(directory, outputFileName);
+}
+
+
 module.exports = {
+    fade,
+    extractClip,
+    truncate,
+    delayAudio,
     getDuration,
     createGIF
 }
